Add deserializer to parse tensor from proto

+++ mace/core/integral_types.h
@@ -9,11 +9,11 @@
 typedef signed char int8;
 typedef short int16;
 typedef int int32;
-typedef long long int64;
+typedef int64_t int64;
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long long uint64;
+typedef uint64_t uint64;
 
 #endif // MACE_CORE_INTEGRAL_TYPES_H_
